moved other utils classes
