+++ src/Util/UrlUtil.php
@@ -6,7 +6,7 @@
  * @license LGPL-3.0-or-later
  */
 
-namespace HeimrichHannot\UtilsBundle\Util\Request;
+namespace HeimrichHannot\UtilsBundle\Util;
 
 use HeimrichHannot\UtilsBundle\Exception\InvalidUrlException;
 use Symfony\Component\HttpFoundation\RequestStack;
